Changed code to drain from read 1 filtered message
Per Brent the setCANfilter call is a hardware level filter which would prevent the RCP from seeing filtered CAN messages throughout the system leading to EMU messages not being seen by the RCP's CAN bus Mapping feature.

This version of the lua script however does not work properly as CAN messages come in too quickly for it to ever exit the repeat until id == nil loop. I think the next step for this is to remove the timeout from rxCAN by setting it to 0 or 1 perhaps.